requirements.txt
numpy==1.19.5
opencv-python-headless
tensorflow-cpu
Flask==1.1.2
gunicorn